+++ frontend/package.json
@@ -31,7 +31,6 @@
     "react-syntax-highlighter": "^15.5.0",
     "tailwind-merge": "^2.2.2",
     "vite": "^5.1.6",
-    "vite-tsconfig-paths": "^4.3.2",
     "web-vitals": "^2.1.4",
     "xterm-addon-fit": "^0.8.0"
   },
@@ -83,6 +82,7 @@
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.2",
     "typescript": "^5.4.3",
+    "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.5.0"
   },
   "packageManager": "npm@10.5.0",
