+++ frontend/src/App.tsx
@@ -2,18 +2,18 @@ import React, { useEffect, useState } from "react";
 import "./App.css";
 import { Toaster } from "react-hot-toast";
 import { useDisclosure } from "@nextui-org/react";
-import CogTooth from "./assets/cog-tooth";
-import ChatInterface from "./components/ChatInterface";
-import Errors from "./components/Errors";
-import { Container, Orientation } from "./components/Resizable";
-import Terminal from "./components/Terminal";
-import Workspace from "./components/Workspace";
-import { fetchMsgTotal } from "./services/session";
-import { initializeAgent } from "./services/settingsService";
-import Socket from "./services/socket";
-import { ResFetchMsgTotal } from "./types/ResponseType";
-import SettingsModal from "./components/modals/settings/SettingsModal";
-import LoadPreviousSessionModal from "./components/modals/load-previous-session/LoadPreviousSessionModal";
+import CogTooth from "#/assets/cog-tooth";
+import ChatInterface from "#/components/ChatInterface";
+import Errors from "#/components/Errors";
+import { Container, Orientation } from "#/components/Resizable";
+import Terminal from "#/components/Terminal";
+import Workspace from "#/components/Workspace";
+import { fetchMsgTotal } from "#/services/session";
+import { initializeAgent } from "#/services/settingsService";
+import Socket from "#/services/socket";
+import { ResFetchMsgTotal } from "#/types/ResponseType";
+import SettingsModal from "#/components/modals/settings/SettingsModal";
+import LoadPreviousSessionModal from "#/components/modals/load-previous-session/LoadPreviousSessionModal";
 
 interface Props {
   setSettingOpen: (isOpen: boolean) => void;
