+++ frontend/src/services/chatService.ts
@@ -4,12 +4,12 @@ import {
   appendUserMessage,
   takeOneTypeIt,
   toggleTypingActive,
-} from "../state/chatSlice";
+} from "#/state/chatSlice";
+import store from "#/store";
+import ActionType from "#/types/ActionType";
+import { SocketMessage } from "#/types/ResponseType";
+import { ActionMessage } from "#/types/Message";
 import Socket from "./socket";
-import store from "../store";
-import ActionType from "../types/ActionType";
-import { SocketMessage } from "../types/ResponseType";
-import { ActionMessage } from "../types/Message";
 
 export function sendChatMessage(message: string): void {
   store.dispatch(appendUserMessage(message));
