+++ frontend/vite.config.ts
@@ -2,6 +2,7 @@
 import { defineConfig, loadEnv } from "vite";
 import react from "@vitejs/plugin-react";
 import viteTsconfigPaths from "vite-tsconfig-paths";
+import path from "path";
 
 export default defineConfig(({ mode }) => {
   const {
